+++ packages/app/imports/ui/components/pre-match/Match.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import Spells from '../Spells';
 import extendMatchResult from '/imports/shared/matches/extendMatchResult/index.ts';
-import perks from '../../../api/riot-api/static/perks';
+import perks from '/imports/shared/riot-api/perks.ts';
 import universeTheme from '../../layouts/universeTheme';
 
 const styles = {
