+++ packages/www/components/CS/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CS';
