+++ packages/www/components/KDA/index.ts
@@ -1,0 +1,1 @@
+export { default } from './KDA';
