+++ packages/www/components/Matches/Matches.tsx
@@ -1,8 +1,20 @@
-import { Typography } from '@material-ui/core';
+import Match from 'components/Match';
 import React, { FunctionComponent } from 'react';
+import { IMatchWithTimeline } from 'shared/riot-api/typings';
 
-const Matches: FunctionComponent = () => {
-  return <Typography>Under construction</Typography>;
+interface MatchesProps {
+  matches: IMatchWithTimeline[];
+  summonerName: string;
+}
+
+const Matches: FunctionComponent<MatchesProps> = ({ matches, summonerName }) => {
+  return (
+    <>
+      {matches.map(match => (
+        <Match key={match.gameId} match={match} summonerName={summonerName} />
+      ))}
+    </>
+  );
 };
 
 export default Matches;
