+++ packages/www/components/Perk/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Perk';
