+++ packages/www/components/Spell/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Spell';
