+++ packages/www/components/Summoner/Summoner.tsx
@@ -1,19 +1,11 @@
-import { Tab, Tabs, Typography } from '@material-ui/core';
+import { Tab, Tabs } from '@material-ui/core';
 import { makeStyles } from '@material-ui/styles';
 import React, { FunctionComponent, useState } from 'react';
-import { ILeaguePositions, ISummoner } from '../../shared/riot-api/typings';
 import Matches from '../Matches';
 import SummonerInfo from '../SummonerInfo';
 import Trees from '../Trees';
 import Trophies from '../Trophies';
-
-interface ISummonerProps {
-  summoner: ISummoner;
-  region: string;
-  summonerName: string;
-  leaguePositions: ILeaguePositions;
-  trophyHunter: any;
-}
+import { ISummonerProps } from './typings';
 
 const useStyles = makeStyles({
   container: {
@@ -36,12 +28,13 @@ const Summoner: FunctionComponent<ISummonerProps> = ({
   region,
   summoner,
   leaguePositions,
-  trophyHunter
+  trophyHunter,
+  lastMatches
 }) => {
   const classes = useStyles();
   const [value, setValue] = useState(0);
 
-  const handleChange = (event, newValue) => {
+  const handleChange = (_event: React.ChangeEvent<{}>, newValue: any) => {
     setValue(newValue);
   };
 
@@ -55,7 +48,7 @@ const Summoner: FunctionComponent<ISummonerProps> = ({
             trophyHunter={trophyHunter}
             {...summoner}
           />
-          <Tabs value={value} onChange={handleChange}>
+          <Tabs value={value} onChange={handleChange} centered>
             <Tab label="Trees" />
             <Tab label="Trophies" />
             <Tab label="Matches" />
@@ -70,7 +63,7 @@ const Summoner: FunctionComponent<ISummonerProps> = ({
             {value === 1 && (
               <Trophies trophiesObtained={trophyHunter && trophyHunter.trophiesObtained} />
             )}
-            {value === 2 && <Matches />}
+            {value === 2 && <Matches matches={lastMatches} summonerName={summoner.name} />}
           </div>
         </div>
       </div>
