+++ packages/www/components/Summoner/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Summoner';
+export * from './typings';
