+++ packages/www/pages/summoner.tsx
@@ -1,17 +1,20 @@
 import './_bootstrap';
 
 import axios from 'axios';
+import Summoner, { ISummonerProps } from 'components/Summoner';
+import Page from 'layouts/Page';
+import { NextFunctionComponent } from 'next';
 import React from 'react';
-import Summoner from '../components/Summoner';
-import Page from '../layouts/Page';
 import {
   getLeaguePositions,
+  getMatchList,
+  getMatchWithTimeline,
   getPlatformIdByRegion,
   getSummoner,
   getTrophyHunter
-} from '../shared/th-api';
+} from 'shared/th-api';
 
-const SummonerPage = props => (
+const SummonerPage: NextFunctionComponent<ISummonerProps> = props => (
   <Page {...props}>
     <Summoner {...props} />{' '}
   </Page>
@@ -19,7 +22,7 @@ const SummonerPage = props => (
 
 SummonerPage.getInitialProps = async ({ query }) => {
   const { region, summonerName } = query;
-  if (!region || !summonerName) {
+  if (!region || typeof region !== 'string' || !summonerName || typeof summonerName !== 'string') {
     throw new Error('invalid query');
   }
 
@@ -28,12 +31,22 @@ SummonerPage.getInitialProps = async ({ query }) => {
     const summonerPromise = getSummoner({ platformId, summonerName });
 
     const trophyHunterPromise = getTrophyHunter({ region, summonerName });
-
     const [summoner, trophyHunter] = await axios.all([summonerPromise, trophyHunterPromise]);
 
-    const leaguePositions = await getLeaguePositions({ platformId, summonerId: summoner.id });
-
-    return { region, summoner, summonerName, leaguePositions, trophyHunter };
+    const leaguePositionsPromise = getLeaguePositions({ platformId, summonerId: summoner.id });
+    const lastMatchesPromise = getMatchList({ platformId, accountId: summoner.accountId }).then(
+      matchList => {
+        const matchPromises = matchList.matches
+          .slice(0, 10)
+          .map(match => getMatchWithTimeline({ platformId, matchId: match.gameId }));
+        return Promise.all(matchPromises);
+      }
+    );
+    const [leaguePositions, lastMatches] = await axios.all([
+      leaguePositionsPromise,
+      lastMatchesPromise
+    ]);
+    return { region, summoner, summonerName, leaguePositions, trophyHunter, lastMatches };
   } catch (error) {
     return {
       region,
@@ -43,7 +56,9 @@ SummonerPage.getInitialProps = async ({ query }) => {
         summonerLevel: 0
       },
       summonerName,
-      leaguePositions: []
+      leaguePositions: [],
+      lastMatches: [],
+      trophyHunter: null
     };
   }
 };
